refactor: 부동소수점 오차 방지를 위해 BigDecimal 사용  #24

+++ src/main/java/com/tavemakers/surf/domain/activity/dto/response/ActivityRecordResDTO.java
@@ -37,8 +37,8 @@ public record ActivityRecordResDTO(
                 .activityName(activityName)
                 .activityDate(formattedDate)
                 .scoreType(activityRecord.getScoreType().name())
-                .prefixSum(activityRecord.getPrefixSum())
-                .appliedScore(activityRecord.getAppliedScore())
+                .prefixSum(activityRecord.getPrefixSum().doubleValue())
+                .appliedScore(activityRecord.getAppliedScore().doubleValue())
                 .build();
     }
